Move stats to separate file
